requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.6.2.post1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
av==14.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.45.5
blinker==1.9.0
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
dotenv==0.9.9
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
faiss-cpu==1.11.0.post1
fastapi==0.115.5
ffmpeg-python==0.2.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.10.0
fugashi==1.4.0
future==1.0.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
grpcio==1.71.0
grpcio-tools==1.71.0
gurobipy==12.0.1
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hdbscan==0.8.40
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
jaconv==0.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.12.0
loguru==0.7.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
medmnist==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
namex==0.0.9
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.4
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==21.10
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.13.3
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.4.5
scikit-image==0.25.1
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.34.2
sentence-transformers==3.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.41.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tqdm==4.67.0
trio==0.30.0
trio-websocket==0.12.2
triton==3.2.0
typing_extensions==4.14.1
tzdata==2025.1
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.32.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
